% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbindrows-package.R
\docType{package}
\name{testbindrows-package}
\alias{testbindrows}
\alias{testbindrows-package}
\title{testbindrows: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
