% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbindrows-package.R
\docType{package}
\name{testbindrows-package}
\alias{testbindrows}
\alias{testbindrows-package}
\title{testbindrows: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/testbindrows}
  \item Report bugs at \url{https://github.com/DavisVaughan/testbindrows/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
